[maven-release-plugin]  copy for tag v5

Conflicts:
	solr/custom/pom.xml
	solr/distr/pom.xml
	solr/pom.xml

+++ service/aas/authentication/src/main/java/eu/europeana/cloud/service/aas/authentication/CassandraAuthenticationService.java
@@ -1,19 +1,21 @@
 package eu.europeana.cloud.service.aas.authentication;
 
+import eu.europeana.cloud.common.model.IdentifierErrorInfo;
 import eu.europeana.cloud.common.model.User;
-import eu.europeana.cloud.service.aas.authentication.exception.*;
+import eu.europeana.cloud.service.aas.authentication.exception.DatabaseConnectionException;
+import eu.europeana.cloud.service.aas.authentication.exception.UserDoesNotExistException;
+import eu.europeana.cloud.service.aas.authentication.exception.UserExistsException;
 import eu.europeana.cloud.service.aas.authentication.repository.CassandraUserDAO;
+import eu.europeana.cloud.service.aas.authentication.status.IdentifierErrorTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 /**
- *
  * Used throughout the Spring Security framework to pass user specific data.
  *
  * @author emmanouil.koufakis@theeuropeanlibrary.org
- *
  */
 public class CassandraAuthenticationService implements UserDetailsService, AuthenticationService {
 
@@ -32,35 +34,67 @@ public class CassandraAuthenticationService implements UserDetailsService, Authe
     public UserDetails loadUserByUsername(final String userName)
             throws UsernameNotFoundException {
         try {
+            SpringUser user = userDao.getUser(userName);
+            if (user == null)
+                throw new UsernameNotFoundException("Username '" + userName + "' could not be retrieved from the database!");
             return userDao.getUser(userName);
         } catch (DatabaseConnectionException ex) {
             throw new UsernameNotFoundException("Username '" + userName + "' could not be retrieved due to database error!", ex);
-        } catch (UserDoesNotExistException ex) {
-            throw new UsernameNotFoundException("Username '" + userName + "' could not be retrieved from the database!", ex);
         }
     }
 
     @Override
     public User getUser(String userName) throws DatabaseConnectionException, UserDoesNotExistException {
+        SpringUser user = userDao.getUser(userName);
+        if (user == null) {
+            throw new UserDoesNotExistException(
+                    new IdentifierErrorInfo(
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getHttpCode(),
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getErrorInfo(userName)));
+        }
         return userDao.getUser(userName);
     }
 
     @Override
-    public void createUser(final User user) throws DatabaseConnectionException, UserExistsException,
-            InvalidUsernameException, InvalidPasswordException {
-        userDao.createUser(user);
+    public void createUser(final User user) throws DatabaseConnectionException, UserExistsException {
+        if (userDao.getUser(user.getUsername()) != null) {
+            throw new UserExistsException(new IdentifierErrorInfo(
+                    IdentifierErrorTemplate.USER_EXISTS.getHttpCode(),
+                    IdentifierErrorTemplate.USER_EXISTS.getErrorInfo(user
+                            .getUsername())));
+        } else {
+            userDao.createUser(user);
+        }
     }
 
     @Override
     public void updateUser(final User user)
-            throws DatabaseConnectionException, UserDoesNotExistException,
-            InvalidPasswordException {
-        userDao.updateUser(user);
+            throws DatabaseConnectionException, UserDoesNotExistException {
+        if (userDao.getUser(user.getUsername()) == null) {
+            throw new UserDoesNotExistException(
+                    new IdentifierErrorInfo(
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getHttpCode(),
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getErrorInfo(user.getUsername())));
+        } else {
+            userDao.updateUser(user);
+        }
     }
 
     @Override
     public void deleteUser(final String userName) throws DatabaseConnectionException, UserDoesNotExistException {
-        userDao.deleteUser(userName);
+        if (userDao.getUser(userName) == null) {
+            throw new UserDoesNotExistException(
+                    new IdentifierErrorInfo(
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getHttpCode(),
+                            IdentifierErrorTemplate.USER_DOES_NOT_EXIST
+                                    .getErrorInfo(userName)));
+        } else {
+            userDao.deleteUser(userName);
+        }
     }
-
 }
